Prime Numbers: * Move StaticPrimes to Primes package + make all methods and class package-private * Move PrimeValidation to Primes package * Update BytePrimeCache - optimize imports * Update ShortPrimeCache - optimize imports
