--- numbers/src/main/java/mathtools/numbers/primes/validate/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Searching for and Validating Prime Numbers
- * @author DK96-OS
- */
-package mathtools.numbers.primes.validate;
